requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.7.2
django-js-asset==3.1.0
sqlparse==0.5.3
